+++ src/main/java/lagatrix/server/connection/communicators/CommunicatorBase.java
@@ -1,10 +1,13 @@
 package lagatrix.server.connection.communicators;
 
 /**
- *
+ * 
+ * 
  * @author javierfh03
  * @since 0.2
  */
 public interface CommunicatorBase {
+    public void obtainRequest();
     
+    public void sendResponse();
 }